// src/models/customers.model.ts
export interface CUSTOMERS  {
    customers_id?:  number;
    customers_gender:  string;
    customers_firstname:  string;
    customers_lastname:  string;
    customers_dob:  string;
    customers_email_address:  string;
    customers_default_address_id:  number;
    customers_telephone:  string;
    customers_fax:  string;
    customers_password:  string;
    customers_newsletter:  string
}